Update switch state naming + update relevanceId prop type + move styled component
